feat: optimize pivot header performance
