+++ common/changes/@visactor/vtable/fix-bug-fix-0.15.4_2023-12-06-08-31.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@visactor/vtable",
+      "comment": "feat: optimize pivot header performance",
+      "type": "none"
+    }
+  ],
+  "packageName": "@visactor/vtable"
+}
